Move local runner to its own namespace
